smartscholar-backend/requirements.txt
fastapi
uvicorn
supabase
pdfplumber
python-dotenv
pydantic
python-multipart
groq
